+++ src/main/java/com/toptal/essienntaemmanuel2ndattempt/domain/Settings.java
@@ -1,6 +1,7 @@
 package com.toptal.essienntaemmanuel2ndattempt.domain;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -19,7 +20,7 @@ public class Settings implements Serializable {
     private Long id;
 
     @Column(name = "expected_num_calories")
-    private Long expectedNumberOfCalories = 10L; // Default
+    private BigDecimal expectedNumberOfCalories = BigDecimal.valueOf(10); // Default
 
     public Settings() {
     }
@@ -32,11 +33,11 @@ public class Settings implements Serializable {
         this.id = id;
     }
 
-    public Long getExpectedNumberOfCalories() {
+    public BigDecimal getExpectedNumberOfCalories() {
         return expectedNumberOfCalories;
     }
 
-    public void setExpectedNumberOfCalories(Long expectedNumberOfCalories) {
+    public void setExpectedNumberOfCalories(BigDecimal expectedNumberOfCalories) {
         this.expectedNumberOfCalories = expectedNumberOfCalories;
     }
