+++ src/main/java/com/toptal/essienntaemmanuel2ndattempt/dto/SettingsDto.java
@@ -3,6 +3,7 @@ package com.toptal.essienntaemmanuel2ndattempt.dto;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * @author bodmas
@@ -13,16 +14,16 @@ public class SettingsDto implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @JsonProperty("expected_num_calories")
-    private Long expectedNumberOfCalories;
+    private BigDecimal expectedNumberOfCalories;
 
     public SettingsDto() {
     }
 
-    public Long getExpectedNumberOfCalories() {
+    public BigDecimal getExpectedNumberOfCalories() {
         return expectedNumberOfCalories;
     }
 
-    public void setExpectedNumberOfCalories(Long expectedNumberOfCalories) {
+    public void setExpectedNumberOfCalories(BigDecimal expectedNumberOfCalories) {
         this.expectedNumberOfCalories = expectedNumberOfCalories;
     }
 }
